refactor: centralize node runtime ownership
