+++ internal/node/node.go
@@ -3,10 +3,8 @@ package node
 import (
 	"context"
 	"encoding/hex"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"net/http"
 	"os"
 	"runtime/debug"
 	"sort"
@@ -14,27 +12,17 @@ import (
 	"time"
 
 	"github.com/LeJamon/go-xrpl/amendment"
-	"github.com/LeJamon/go-xrpl/codec/addresscodec"
 	"github.com/LeJamon/go-xrpl/config"
 	"github.com/LeJamon/go-xrpl/drops"
 	"github.com/LeJamon/go-xrpl/internal/consensus"
 	"github.com/LeJamon/go-xrpl/internal/consensus/adaptor"
-	"github.com/LeJamon/go-xrpl/internal/ledger/cleaner"
 	"github.com/LeJamon/go-xrpl/internal/ledger/genesis"
 	"github.com/LeJamon/go-xrpl/internal/ledger/service"
-	"github.com/LeJamon/go-xrpl/internal/ledger/shamapstore"
-	"github.com/LeJamon/go-xrpl/internal/manifest"
-	"github.com/LeJamon/go-xrpl/internal/observability"
 	"github.com/LeJamon/go-xrpl/internal/peermanagement"
 	"github.com/LeJamon/go-xrpl/internal/peermanagement/message"
 	"github.com/LeJamon/go-xrpl/internal/rpc"
-	"github.com/LeJamon/go-xrpl/internal/rpc/handlers"
-	"github.com/LeJamon/go-xrpl/internal/rpc/loadtrack"
 	"github.com/LeJamon/go-xrpl/internal/rpc/types"
-	validatorlist "github.com/LeJamon/go-xrpl/internal/validator/list"
-	"github.com/LeJamon/go-xrpl/internal/watchdog"
 	xrpllog "github.com/LeJamon/go-xrpl/log"
-	"github.com/LeJamon/go-xrpl/shamap/backend"
 	kvpebble "github.com/LeJamon/go-xrpl/storage/kvstore/pebble"
 	"github.com/LeJamon/go-xrpl/storage/nodestore"
 	"github.com/LeJamon/go-xrpl/storage/relationaldb"
@@ -173,980 +161,17 @@ func run(
 	startup service.StartupConfig,
 	rootLogger, serverLog xrpllog.Logger,
 	options RunOptions,
-) (runErr error) {
-	ctx, cancelRuntime := context.WithCancelCause(ctx)
-	defer cancelRuntime(nil)
-	if err := context.Cause(ctx); err != nil {
-		return err
-	}
-	if err := validateTrustedValidatorConfig(appConfig, standalone); err != nil {
-		return err
-	}
-
-	var err error
-	// Pre-declared so the deferred shutdown can clean up whatever the
-	// init path managed to populate before any error return. doShutdown
-	// tolerates nil components for the partial-init case.
-	var (
-		db                  nodestore.Database
-		repoManager         relationaldb.RepositoryManager
-		ledgerService       *service.Service
-		ledgerCleaner       *cleaner.Cleaner
-		consensusComponents *adaptor.Components
-		rotator             *shamapstore.Rotator
-		transports          *boundRPCTransports
-		wsServer            *rpc.WebSocketServer
-		stallWatchdog       *watchdog.Watchdog
-	)
-	defer func() {
-		cancelRuntime(nil)
-		if options.Stopping != nil {
-			options.Stopping()
-		}
-		if shutdownErr := doShutdown(transports, wsServer, ledgerService, ledgerCleaner, consensusComponents, rotator, db, repoManager, serverLog); shutdownErr != nil {
-			runErr = errors.Join(runErr, shutdownErr)
-		}
-	}()
-
-	db, repoManager, err = setupStorage(ctx, appConfig, serverLog)
-	if err != nil {
-		return err
-	}
-	if err := context.Cause(ctx); err != nil {
-		return err
-	}
-	var nodeFamily *backend.NodeStore
-	if db != nil {
-		nodeFamily = backend.New(db)
-	}
-
-	// Load genesis configuration from config file path (if set)
-	genesisFile := appConfig.GenesisFile
-	var genesisConfig genesis.Config
-	if genesisFile != "" {
-		genesisJSON, err := config.LoadGenesisJSON(genesisFile)
-		if err != nil {
-			return fmt.Errorf("load genesis file %q: %w", genesisFile, err)
-		}
-		if err := genesisJSON.Validate(); err != nil {
-			return fmt.Errorf("invalid genesis file %q: %w", genesisFile, err)
-		}
-		genesisCfg, err := genesisJSON.ToGenesisConfig()
-		if err != nil {
-			return fmt.Errorf("parse genesis configuration %q: %w", genesisFile, err)
-		}
-		genesisConfig = genesis.Config{
-			TotalXRP:            genesisCfg.TotalXRP,
-			CloseTimeResolution: genesisCfg.CloseTimeResolution,
-			Fees: genesis.DefaultFees{
-				BaseFee:          genesisCfg.BaseFee,
-				ReserveBase:      genesisCfg.ReserveBase,
-				ReserveIncrement: genesisCfg.ReserveIncrement,
-			},
-			Amendments: genesisCfg.Amendments,
-		}
-		for _, acc := range genesisCfg.InitialAccounts {
-			genesisConfig.InitialAccounts = append(genesisConfig.InitialAccounts, genesis.InitialAccount{
-				Address:  acc.Address,
-				Balance:  acc.Balance,
-				Sequence: acc.Sequence,
-				Flags:    acc.Flags,
-			})
-		}
-		serverLog.Info("Genesis config loaded", "path", genesisFile)
-	} else {
-		genesisConfig = genesis.DefaultConfig()
-		if appConfig.GenesisAmendmentsDisabled {
-			genesisConfig.Amendments = nil
-		}
-		serverLog.Info("Genesis config using built-in defaults")
-	}
-
-	// Get network ID from config
-	networkID, err := appConfig.ResolvedNetworkID()
-	if err != nil {
-		return fmt.Errorf("get network ID: %w", err)
-	}
-
-	// Build the live amendment table from the operator's [amendments] config.
-	// One instance is shared between the ledger service (which folds validated
-	// flag ledgers into it) and the consensus adaptor (which sources vote
-	// stances from it).
-	amendmentTable := buildTable(ctx, appConfig.Amendments, repoManager, serverLog)
-	if err := context.Cause(ctx); err != nil {
-		return err
-	}
-
-	// Build the transaction-queue config from the operator's
-	// [transaction_queue] stanza layered over the rippled defaults.
-	txqCfg, err := service.TxQConfigFromTuning(appConfig.TransactionQueue, standalone)
-	if err != nil {
-		return err
-	}
-
-	// Initialize ledger service
-	cfg := service.Config{
-		Standalone:      standalone,
-		NodeSize:        appConfig.NodeSize,
-		FetchDepth:      effectivePeerFetchDepth(appConfig.GetFetchDepthUint32(), appConfig.NodeDB.OnlineDelete),
-		NetworkID:       uint32(networkID),
-		NodeStore:       db,
-		SHAMapFamily:    nodeFamily,
-		FastLoad:        appConfig.NodeDB.FastLoad,
-		FastLoadWorkers: appConfig.NodeDB.FastLoadWorkers,
-		RelationalDB:    repoManager,
-		Logger:          rootLogger,
-		Table:           amendmentTable,
-		TxQ:             &txqCfg,
-		Startup:         startup,
-	}
-	cfg.GenesisConfig = genesisConfig
-	configuredFees := configuredLedgerLoadFees(appConfig)
-	cfg.ConfiguredFees = &configuredFees
-
-	ledgerService, err = service.New(cfg)
-	if err != nil {
-		return fmt.Errorf("create ledger service: %w", err)
-	}
-
-	if err := ledgerService.Start(); err != nil {
-		return fmt.Errorf("start ledger service: %w", err)
-	}
-	if err := context.Cause(ctx); err != nil {
-		return err
-	}
-
-	sampler := observability.NewSchedLatencySampler()
-	if err := sampler.Start(ctx); err != nil {
-		return fmt.Errorf("start scheduler latency sampler: %w", err)
-	}
-	defer sampler.Stop()
-
-	// Wire up RPC services
-	ledgerAdapter := rpc.NewLedgerServiceAdapter(ledgerService)
-	services := types.NewServiceContainer(ledgerAdapter)
-	services.ServerInfoConfig = serverInfoConfigSnapshot(appConfig)
-
-	// Gate the beta RPC API (api_version 3) on the operator's beta_rpc_api
-	// knob, mirroring rippled Config::BETA_RPC_API.
-	services.BetaRPCAPI = appConfig.BetaRPCAPI != 0
-
-	// Advisory-delete state (can_delete RPC). Available in both standalone and
-	// consensus modes; gated by node_db advisory_delete and persisted under
-	// database_path. Mirrors rippled's SHAMapStore advisory-delete state.
-	if advisoryStore, asErr := shamapstore.New(
-		appConfig.NodeDB.IsOnlineDeleteEnabled() && appConfig.NodeDB.IsAdvisoryDeleteEnabled(),
-		appConfig.LocalStateDir(),
-	); asErr != nil {
-		if appConfig.NodeDB.IsOnlineDeleteEnabled() {
-			return fmt.Errorf("load online-delete state: %w", asErr)
-		}
-		serverLog.Warn("Failed to load advisory-delete state", "err", asErr)
-	} else {
-		services.AdvisoryDeleteState = advisoryStore
-
-		// Online-delete rotation: when node_db online_delete is set and the
-		// node store can enumerate its keyspace, run a background job that
-		// reclaims disk by deleting complete ledgers below the rotation
-		// boundary. NewRotator returns nil when online_delete is off.
-		if appConfig.NodeDB.IsOnlineDeleteEnabled() {
-			if prunable, ok := db.(shamapstore.NodePruner); ok {
-				var relPruner shamapstore.RelationalPruner
-				if repoManager != nil {
-					relPruner = relationaldb.NewLedgerPruner(repoManager, appConfig.NodeDB.DeleteBatch)
-				}
-				rotator = shamapstore.NewRotator(
-					advisoryStore,
-					prunable,
-					relPruner,
-					shamapstore.RotationConfig{
-						DeleteInterval: uint32(appConfig.NodeDB.OnlineDelete),
-						DeleteBatch:    appConfig.NodeDB.DeleteBatch,
-						BackOff:        time.Duration(appConfig.NodeDB.BackOffMilliseconds) * time.Millisecond,
-					},
-					serverLog,
-				)
-				if err := rotator.ReconcileGenerationState(); err != nil {
-					return fmt.Errorf("reconcile online-delete generation state: %w", err)
-				}
-				minimumOnline := rotator.MinimumOnline()
-				if minimumOnline == 0 && repoManager != nil {
-					minSeq, minErr := repoManager.Ledger().GetMinLedgerSeq(ctx)
-					if minErr != nil {
-						return fmt.Errorf("load online-delete minimum ledger: %w", minErr)
-					}
-					if minSeq != nil {
-						minimumOnline = uint32(*minSeq)
-						if err := rotator.SetMinimumOnlineFloor(minimumOnline); err != nil {
-							return fmt.Errorf("persist online-delete minimum ledger: %w", err)
-						}
-					}
-				}
-				if minimumOnline > 0 {
-					nodeFamily.SetMinimumLedgerSeq(minimumOnline)
-				}
-				rotator.SetStateRefresh(
-					ledgerService.RefreshValidatedState,
-					nodeFamily.SetMinimumLedgerSeq,
-					nodeFamily.BeginPrune,
-				)
-				if err := context.Cause(ctx); err != nil {
-					return err
-				}
-				rotator.Start()
-				serverLog.Info("Online delete enabled",
-					"online_delete", appConfig.NodeDB.OnlineDelete,
-					"advisory_delete", appConfig.NodeDB.IsAdvisoryDeleteEnabled())
-			} else {
-				serverLog.Warn("online_delete configured but node store backend does not support pruning")
-			}
-		}
-	}
-	retentionFloor := func() uint32 {
-		floor := uint32(appConfig.NodeDB.EarliestSeq)
-		if rotator != nil && rotator.MinimumOnline() > floor {
-			floor = rotator.MinimumOnline()
-		}
-		return floor
-	}
-	if appConfig.NodeDB.EarliestSeq > 0 || rotator != nil {
-		ledgerService.SetMinimumOnlineFunc(retentionFloor)
-		if nodeFamily != nil {
-			nodeFamily.SetMinimumLedgerSeq(retentionFloor())
-		}
-	}
-
-	// TxQ metrics are available in both standalone and consensus modes,
-	// so wire the server_info hook before the !standalone branch.
-	ledgerSvcRef := ledgerService
-	services.TxQMetrics = func() types.TxQServerMetrics {
-		m := ledgerSvcRef.TxQMetrics()
-		return types.TxQServerMetrics{
-			ReferenceFeeLevel:     m.ReferenceFeeLevel,
-			MinProcessingFeeLevel: m.MinProcessingFeeLevel,
-			OpenLedgerFeeLevel:    m.OpenLedgerFeeLevel,
-		}
-	}
-	services.TxQFeeMetrics = func() types.TxQFeeMetrics {
-		m := ledgerSvcRef.TxQMetrics()
-		return types.TxQFeeMetrics{
-			TxCount:               m.TxCount,
-			TxQMaxSize:            m.TxQMaxSize,
-			TxInLedger:            m.TxInLedger,
-			TxPerLedger:           m.TxPerLedger,
-			ReferenceFeeLevel:     m.ReferenceFeeLevel,
-			MinProcessingFeeLevel: m.MinProcessingFeeLevel,
-			MedFeeLevel:           m.MedFeeLevel,
-			OpenLedgerFeeLevel:    m.OpenLedgerFeeLevel,
-		}
-	}
-	services.QueueAccountTxs = func(account [20]byte) []types.QueuedTxInfo {
-		return queuedTxInfos(ledgerSvcRef.QueueAccountTxs(account))
-	}
-	services.QueueAllTxs = func() []types.QueuedTxInfo {
-		return queuedTxInfos(ledgerSvcRef.QueueAllTxs())
-	}
-
-	// get_counts surfaces node-store I/O counters and locally-held
-	// transactions. Available in both standalone and consensus modes since it
-	// only needs the ledger service.
-	services.GetCounts = func() types.CountsResult {
-		c := ledgerSvcRef.Counts()
-		res := types.CountsResult{
-			Standalone: c.Standalone,
-			LocalTxs:   c.LocalTxs,
-		}
-		if c.NodeStore != nil {
-			res.NodeStore = &types.NodeStoreCounts{
-				Reads:      c.NodeStore.Reads,
-				FetchHits:  c.NodeStore.FetchHits,
-				Writes:     c.NodeStore.Writes,
-				ReadBytes:  c.NodeStore.ReadBytes,
-				WriteBytes: c.NodeStore.WriteBytes,
-			}
-		}
-		if c.FullBelow != nil {
-			res.FullBelow = &types.FullBelowCounts{
-				Size: c.FullBelow.Size, TargetSize: c.FullBelow.TargetSize,
-				Hits: c.FullBelow.Hits, Misses: c.FullBelow.Misses,
-				Inserts: c.FullBelow.Inserts, Evictions: c.FullBelow.Evictions,
-				Sweeps: c.FullBelow.Sweeps,
-			}
-		}
-		return res
-	}
-
-	// LoadFactorFees surfaces the local/net/cluster fee factors that
-	// drive the admin-only human-mode load_factor_local / load_factor_net /
-	// load_factor_cluster emissions (NetworkOPs.cpp:2887-2901). Net here
-	// mirrors rippled's "remote" axis — LoadFeeTrack stores it under
-	// remoteFee_. The closure re-reads on every server_info call so the
-	// hook tracks live tracker state without rewiring.
-	services.LoadFactorFees = func() types.LoadFactorFees {
-		ft := ledgerSvcRef.FeeTrack()
-		if ft == nil {
-			base := uint32(256)
-			return types.LoadFactorFees{Local: base, Net: base, Cluster: base}
-		}
-		return types.LoadFactorFees{
-			Local:   ft.LocalFee(),
-			Net:     ft.RemoteFee(),
-			Cluster: ft.ClusterFee(),
-		}
-	}
-	services.IsLoadedCluster = func() bool {
-		ft := ledgerSvcRef.FeeTrack()
-		return ft != nil && ft.IsLoadedCluster()
-	}
-
-	// Background ledger-integrity verification requires the same durable SHAMap
-	// family used by the ledger service. A private in-memory fallback cannot
-	// verify persisted ledger contents and would report misleading failures.
-	var cleanerSource *ledgerCleanerSource
-	if nodeFamily != nil {
-		cleanerSource = &ledgerCleanerSource{svc: ledgerSvcRef, family: nodeFamily}
-		ledgerCleaner = cleaner.New(cleanerSource, rootLogger)
-		if err := context.Cause(ctx); err != nil {
-			return err
-		}
-		ledgerCleaner.Start()
-
-		cleanerRef := ledgerCleaner
-		services.LedgerCleanerConfigure = func(p types.LedgerCleanerParams) types.LedgerCleanerStatus {
-			return toCleanerStatus(cleanerRef.Clean(cleaner.Params{
-				Ledger:     p.Ledger,
-				MinLedger:  p.MinLedger,
-				MaxLedger:  p.MaxLedger,
-				Full:       p.Full,
-				CheckNodes: p.CheckNodes,
-				FixTxns:    p.FixTxns,
-				Stop:       p.Stop,
-			}))
-		}
-		services.LedgerCleanerStatusFn = func() types.LedgerCleanerStatus {
-			return toCleanerStatus(cleanerRef.Status())
-		}
-	}
-
-	// Start consensus/networking if not in standalone mode
-	if !standalone {
-		var compErr error
-		var validationRepo relationaldb.ValidationRepository
-		if repoManager != nil {
-			validationRepo = repoManager.Validation()
-		}
-		// Pass the online-delete floor to consensus so acquisition and
-		// peer-serving refuse ledgers below the deletion boundary. Keep the
-		// interface nil when rotation is off so the disabled path is unchanged
-		// (a typed-nil *Rotator would be a non-nil interface).
-		var floor adaptor.MinimumOnlineFloor
-		if appConfig.NodeDB.EarliestSeq > 0 || rotator != nil {
-			floor = minimumOnlineFloorFunc(retentionFloor)
-		}
-		consensusComponents, compErr = adaptor.NewFromConfig(ctx, appConfig, ledgerService, validationRepo, floor)
-		if compErr != nil {
-			return fmt.Errorf("create consensus components: %w", compErr)
-		}
-		if rotator != nil {
-			ageThreshold := 60 * time.Second
-			if appConfig.NodeDB.AgeThresholdSeconds > 0 {
-				ageThreshold = time.Duration(appConfig.NodeDB.AgeThresholdSeconds) * time.Second
-			}
-			recoveryWait := 5 * time.Second
-			if appConfig.NodeDB.RecoveryWaitSeconds > 0 {
-				recoveryWait = time.Duration(appConfig.NodeDB.RecoveryWaitSeconds) * time.Second
-			}
-			rotator.SetHealthCheck(func() bool {
-				if consensusComponents.Adaptor.GetOperatingMode() != consensus.OpModeFull {
-					return false
-				}
-				validated := ledgerService.GetValidatedLedger()
-				return validated != nil && time.Since(validated.CloseTime()) <= ageThreshold
-			}, recoveryWait)
-		}
-
-		// Back inbound acquisitions with the node store before Start launches the
-		// router loop, so the family is published before any acquisition reads it
-		// (issue #1158).
-		if nodeFamily != nil {
-			if router := consensusComponents.Router; router != nil {
-				router.SetAcquisitionFamily(nodeFamily)
-			}
-		}
-
-		if router := consensusComponents.Router; router != nil && cleanerSource != nil {
-			cleanerSource.SetReacquire(func(ctx context.Context, seq uint32) error {
-				if err := ctx.Err(); err != nil {
-					return err
-				}
-				_, started, _ := router.RequestLedger([32]byte{}, seq)
-				if !started {
-					return fmt.Errorf("ledger_cleaner: unable to acquire ledger %d", seq)
-				}
-				return nil
-			})
-		}
-
-		// Wire transaction relay: when a tx is submitted via RPC,
-		// broadcast it to peers. LocalTxs holding is handled inside
-		// service.SubmitTransaction so the broadcaster only relays.
-		overlay := consensusComponents.Overlay
-
-		// Closed-Ledger / Previous-Ledger hints (Handshake.cpp:219-223).
-		overlay.SetLedgerHintProvider(func() (peermanagement.LedgerHints, bool) {
-			cl := ledgerService.GetClosedLedger()
-			if cl == nil {
-				return peermanagement.LedgerHints{}, false
-			}
-			return peermanagement.LedgerHints{Closed: cl.Hash(), Parent: cl.ParentHash()}, true
-		})
-
-		overlay.SetValidLedgerProvider(func() (uint32, time.Duration, bool) {
-			vl := ledgerService.GetValidatedLedger()
-			if vl == nil {
-				return 0, 0, false
-			}
-			age := time.Since(vl.CloseTime())
-			return vl.Sequence(), age, true
-		})
-		broadcastTx := newTxBroadcaster(overlay)
-		ledgerAdapter.SetTxBroadcaster(broadcastTx)
-		// Wire OpenLedger.Accept's relay callback so recovered txs are
-		// re-broadcast post-LCL (rippled OpenLedger.cpp:120-150).
-		ledgerService.SetTxRelay(broadcastTx)
-
-		// Wire the open-ledger tx lookup used by the tx-reduce-relay
-		// reply path (TMGetObjectByHash{otTRANSACTIONS} → TMTransactions
-		// reply) and the periodic TMHaveTransactions announce.
-		// Feature-gated downstream by Config.EnableTxReduceRelay; the
-		// providers themselves are always wired so a flip of the
-		// config flag doesn't require a restart-and-rewire.
-		overlay.SetTxProvider(ledgerService.OpenLedgerGetTx)
-		overlay.SetOpenLedgerHashesProvider(ledgerService.OpenLedgerTxHashes)
-
-		// Wire the generic node-object lookup used by the
-		// TMGetObjectByHash by-hash serve path (PeerImp.cpp:2483-2538).
-		// Only wired when a node store is configured; an in-memory
-		// deployment leaves the provider nil and the serve path drops
-		// the request without charging.
-		if db != nil {
-			overlay.SetNodeObjectProvider(func(hash [32]byte) ([]byte, bool) {
-				node, err := db.Fetch(ctx, nodestore.Hash256(hash))
-				if err != nil || node == nil {
-					return nil, false
-				}
-				return node.Data, true
-			})
-		}
-
-		// LoadFeeTrack ingress + outbound self-load advertisement.
-		// Mirrors the rippled wiring split:
-		//   - PeerImp.cpp:1193 setClusterFee(median) on inbound TMCluster
-		//   - NetworkOPs.cpp:1126-1132 self-entry sources getLocalFee()
-		if ft := ledgerSvcRef.FeeTrack(); ft != nil {
-			overlay.SetClusterFeeSink(ft.SetClusterFee)
-			overlay.SetLocalLoadFeeProvider(ft.LocalFee)
-		}
-
-		// Expose node identity and consensus stats to RPC handlers.
-		services.NodePublicKey = consensusComponents.Overlay.Identity().EncodedPublicKey()
-		engine := consensusComponents.Engine
-		services.LastCloseInfo = func() (int, int) {
-			proposers, convergeTime := engine.GetLastCloseInfo()
-			return proposers, int(convergeTime.Milliseconds())
-		}
-		// Expose live consensus-round state to the `consensus_info` RPC
-		// (rippled NetworkOPs::getConsensusInfo → RCLConsensus::getJson).
-		services.ConsensusInfo = engine.GetJSON
-		// Expose the live consensus quorum to the `server_info` RPC so
-		// operators see the actual quorum (recomputed by the adaptor
-		// from UNL ∖ negative-UNL) instead of the hardcoded "1" that
-		// the bootstrap-time field used to return — #451.
-		services.ValidationQuorum = consensusComponents.Adaptor.GetQuorum
-
-		// Peer-disconnect counters and the operating-mode state-accounting
-		// snapshot need the overlay/adaptor, so they live inside the
-		// !standalone branch. (TxQMetrics is wired above; it only needs
-		// the ledger service.)
-		overlayRef := consensusComponents.Overlay
-		services.PeerDisconnects = func() (uint64, uint64) {
-			return overlayRef.PeerDisconnects(), overlayRef.PeerDisconnectsResources()
-		}
-		// jq_trans_overflow folds the two sequential stages where a
-		// saturated inbound transaction is shed: the overlay ingress gate
-		// (max_transactions ceiling) and the consensus worker pool
-		// (Router.DroppedTxJobs). A frame is shed by at most one stage, so
-		// summing the disjoint counts reports the total without double-counting
-		// and mirrors rippled's single jq_trans_overflow counter.
-		routerRef := consensusComponents.Router
-		services.JqTransOverflow = func() uint64 {
-			n := overlayRef.DroppedTransactions()
-			if routerRef != nil {
-				n += routerRef.DroppedTxJobs()
-			}
-			return n
-		}
-		services.TxReduceRelayMetrics = func() types.TxReduceRelayMetrics {
-			s := overlayRef.TxMetricsSnapshot()
-			return types.TxReduceRelayMetrics{
-				TxCnt:           s.TxCnt,
-				TxSz:            s.TxSz,
-				HaveTxCnt:       s.HaveTxCnt,
-				HaveTxSz:        s.HaveTxSz,
-				GetLedgerCnt:    s.GetLedgerCnt,
-				GetLedgerSz:     s.GetLedgerSz,
-				LedgerDataCnt:   s.LedgerDataCnt,
-				LedgerDataSz:    s.LedgerDataSz,
-				TransactionsCnt: s.TransactionsCnt,
-				TransactionsSz:  s.TransactionsSz,
-				SelectedCnt:     s.SelectedCnt,
-				SuppressedCnt:   s.SuppressedCnt,
-				NotEnabledCnt:   s.NotEnabledCnt,
-				MissingTxFreq:   s.MissingTxFreq,
-			}
-		}
-		// Expose the overlay's peer-reservation table to the admin
-		// peer_reservations_* RPCs (nil when no data dir is configured).
-		if reservations := overlayRef.PeerReservations(); reservations != nil {
-			services.PeerReservationAdd = func(nodePublic, description string) (string, bool, error) {
-				prev, err := reservations.Insert(&peermanagement.PeerReservation{NodeID: nodePublic, Description: description})
-				if prev != nil {
-					return prev.Description, true, err
-				}
-				return "", false, err
-			}
-			services.PeerReservationDel = func(nodePublic string) (string, bool, error) {
-				prev, err := reservations.Erase(nodePublic)
-				if prev != nil {
-					return prev.Description, true, err
-				}
-				return "", false, err
-			}
-			services.PeerReservationList = func() []types.PeerReservationEntry {
-				list := reservations.List()
-				out := make([]types.PeerReservationEntry, 0, len(list))
-				for _, r := range list {
-					out = append(out, types.PeerReservationEntry{NodePublic: r.NodeID, Description: r.Description})
-				}
-				return out
-			}
-		}
-		services.PeerConnect = overlayRef.Connect
-		services.ResourceBlacklist = overlayRef.BlacklistJSON
-		acctRef := consensusComponents.Adaptor
-		services.StateAccounting = func() types.StateAccountingSnapshot {
-			snap := acctRef.StateAccounting()
-			if len(snap.Modes) == 0 {
-				return types.StateAccountingSnapshot{}
-			}
-			modes := make(map[string]types.StateAccountingEntry, len(snap.Modes))
-			for mode, entry := range snap.Modes {
-				modes[mode] = types.StateAccountingEntry{
-					Transitions: entry.Transitions,
-					DurationUs:  entry.DurationUs,
-				}
-			}
-			return types.StateAccountingSnapshot{
-				Modes:             modes,
-				CurrentDurationUs: snap.CurrentDurationUs,
-				InitialSyncUs:     snap.InitialSyncUs,
-			}
-		}
-		services.CloseTimeOffset = acctRef.CloseOffset
-
-		// Expose the router's inbound-ledger acquisition tracker to the
-		// fetch_info RPC (rippled InboundLedgers). Populated by the live
-		// sync path; empty until the node is actively acquiring.
-		if router := consensusComponents.Router; router != nil {
-			services.FetchInfo = router.FetchInfo
-			services.FetchInfoClear = router.ClearFetchInfo
-			services.RequestLedger = router.RequestLedger
-			services.FetchPackCacheSize = router.FetchPackCacheSize
-		}
-
-		// Expose the validator-manifest cache to the `manifest` RPC.
-		// The cache is shared — the router writes inbound manifests,
-		// the engine reads for ephemeral→master translation, and this
-		// RPC reads for external queries.
-		services.Manifests = consensusComponents.ValidatorManifests
-
-		// Expose the publisher-list aggregator (when configured) to
-		// the `validators` and `validator_list_sites` RPC methods.
-		// nil-safe: NewRPCReader returns an inert reader when the
-		// aggregator is nil, so the handlers return empty arrays in
-		// that case rather than panicking.
-		services.ValidatorList = validatorlist.NewRPCReader(consensusComponents.ValidatorList)
-
-		// Surface UNL-blocked state (validator list expired) so conditionMet
-		// can return rpcEXPIRED_VALIDATOR_LIST, mirroring rippled's
-		// NetworkOPs::isUNLBlocked. Only when a publisher list is configured.
-		if consensusComponents.ValidatorList != nil {
-			services.UNLBlocked = consensusComponents.ValidatorList.IsUNLBlocked
-		}
-
-		// Expose static config validators, cached signing keys, and the
-		// negative-UNL set to the `validators` RPC so it returns the
-		// same shape rippled's ValidatorList::getJson does.
-		//
-		// Bind to the live accessor (not a boot-time copy) so a SIGHUP
-		// reload of the [validators] stanza is visible to the RPC.
-		componentsRef := consensusComponents
-		adaptorRef := consensusComponents.Adaptor
-		services.LocalStaticTrustedKeysBase58 = func() []string {
-			masters := componentsRef.StaticTrustedMasterKeys()
-			out := make([]string, 0, len(masters))
-			for _, mk := range masters {
-				if enc, err := addresscodec.EncodeNodePublicKey(mk[:]); err == nil {
-					out = append(out, enc)
-				}
-			}
-			return out
-		}
-		services.TrustedValidatorKeysBase58 = func() []string {
-			masters := adaptorRef.GetTrustedMasterKeys()
-			out := make([]string, 0, len(masters))
-			for _, mk := range masters {
-				if enc, err := addresscodec.EncodeNodePublicKey(mk[:]); err == nil {
-					out = append(out, enc)
-				}
-			}
-			return out
-		}
-		if mc := consensusComponents.ValidatorManifests; mc != nil {
-			// Mirrors rippled getJson at ValidatorList.cpp:1726-1734 —
-			// `signing_keys` only surfaces master→signing pairs for
-			// masters present in keyListings_, i.e. validators listed
-			// by at least one publisher, pinned in the local
-			// [validators] stanza, or used as the local identity. Without
-			// this filter we would leak every gossiped manifest, including
-			// ones unrelated to any trusted publisher.
-			vlAgg := consensusComponents.ValidatorList
-			services.SigningKeysBase58 = func() map[string]string {
-				snap := mc.MasterToSigning()
-				if len(snap) == 0 {
-					return nil
-				}
-				listed := make(map[[33]byte]struct{})
-				for _, mk := range adaptorRef.GetTrustedMasterKeys() {
-					listed[mk] = struct{}{}
-				}
-				if vlAgg != nil {
-					for _, p := range vlAgg.PublisherSnapshot() {
-						if p.Status != validatorlist.StatusAvailable {
-							continue
-						}
-						for _, mk := range p.Validators {
-							listed[mk] = struct{}{}
-						}
-					}
-				}
-				if len(listed) == 0 {
-					return nil
-				}
-				out := make(map[string]string, len(listed))
-				for master, signing := range snap {
-					if _, ok := listed[master]; !ok {
-						continue
-					}
-					mEnc, mErr := addresscodec.EncodeNodePublicKey(master[:])
-					sEnc, sErr := addresscodec.EncodeNodePublicKey(signing[:])
-					if mErr == nil && sErr == nil {
-						out[mEnc] = sEnc
-					}
-				}
-				return out
-			}
-		}
-		services.NegativeUNLBase58 = func() []string {
-			masters := adaptorRef.GetNegativeUNLMasters()
-			if len(masters) == 0 {
-				return nil
-			}
-			out := make([]string, 0, len(masters))
-			for _, mk := range masters {
-				if enc, err := addresscodec.EncodeNodePublicKey(mk[:]); err == nil {
-					out = append(out, enc)
-				}
-			}
-			return out
-		}
-
-		// Expose the local validator's 33-byte signing public key to
-		// validator_info / server_info. Mirrors rippled's
-		// getValidationPublicKey gate: empty means the server is not
-		// configured as a validator and the handlers return "not a
-		// validator" / "none". GetValidatorKey returns the 20-byte
-		// NodeID, NOT the public key — copying it into a 33-byte slice
-		// zero-padded the last 13 bytes and produced a bogus key.
-		if pk, err := consensusComponents.Adaptor.GetValidatorSigningKey(); err == nil {
-			services.ValidatorPublicKey = append([]byte(nil), pk[:]...)
-		}
-
-		isValidator := appConfig.IsValidator()
-		serverLog.Info("Running in consensus mode",
-			"validator", isValidator,
-			"peers", len(appConfig.IPs)+len(appConfig.IPsFixed),
-		)
-	} else {
-		genesisAddr, _ := ledgerService.GetGenesisAccount()
-		serverLog.Info("Running in standalone mode",
-			"genesisAccount", genesisAddr,
-			"validatedLedger", ledgerService.GetValidatedLedgerIndex(),
-			"openLedger", ledgerService.GetCurrentLedgerIndex(),
-		)
-	}
-
-	transportLoad := loadtrack.New()
-
-	// Create HTTP JSON-RPC server. The dispatch timeout stays strictly below
-	// the transport WriteTimeout (see httpWriteTimeout) so a timed-out request
-	// can still serialize its error envelope.
-	httpServer := rpc.NewServerWithLoadTracker(rpcDispatchTimeout, services, transportLoad)
-	if consensusComponents != nil && consensusComponents.Overlay != nil {
-		httpServer.SetPeerSource(consensusComponents.Overlay)
-	}
-
-	services.SetDispatcher(httpServer)
-
-	// Create WebSocket server for real-time subscriptions
-	wsServer = rpc.NewWebSocketServerWithLoadTracker(rpcDispatchTimeout, services, transportLoad)
-	if appConfig.WebsocketPingFrequency > 0 {
-		wsServer.SetPingInterval(time.Duration(appConfig.WebsocketPingFrequency) * time.Second)
-	}
-	wsServer.RegisterAllMethods()
-	if consensusComponents != nil && consensusComponents.Overlay != nil {
-		wsServer.SetPeerSource(consensusComponents.Overlay)
-	}
-
-	// Create a ledger info provider adapter for WebSocket subscribe responses
-	wsServer.SetLedgerInfoProvider(&ledgerInfoAdapter{ledgerService: ledgerService})
-
-	publisher := rpc.NewPublisher(wsServer.SubscriptionManager())
-
-	// Wire the WebSocket event sources that previously had a publisher
-	// helper but no upstream subscriber. Each call mirrors a rippled
-	// pubXxx feed (NetworkOPs.cpp); without them the corresponding
-	// streams accepted subscribers but never delivered.
-	if consensusComponents != nil && consensusComponents.Overlay != nil {
-		// pubPeerStatus → peer_status (NetworkOPs.cpp:2514-2540).
-		consensusComponents.Overlay.SetPeerStatusPublisher(func(u peermanagement.PeerStatusUpdate) {
-			publisher.PublishPeerStatus(&rpc.PeerStatusEvent{
-				Type:           "peerStatusChange",
-				Status:         u.Status,
-				Action:         u.Action,
-				Date:           u.Date,
-				LedgerHash:     u.LedgerHash,
-				LedgerIndex:    u.LedgerIndex,
-				LedgerIndexMin: u.LedgerIndexMin,
-				LedgerIndexMax: u.LedgerIndexMax,
-			})
-		})
-
-		// pubManifest → manifests (NetworkOPs.cpp:2234-2261). One sink
-		// installed on the cache, fed by every accepted manifest
-		// regardless of source (overlay relay, startup, validator-list
-		// aggregator, local-manifest emit).
-		if consensusComponents.ValidatorManifests != nil {
-			consensusComponents.ValidatorManifests.SubscribeAccepted(func(m *manifest.Manifest) {
-				publishManifestIfSubscribed(publisher, m)
-			})
-		}
-
-		// pubValidation + pubConsensus → validations / consensus
-		// (NetworkOPs.cpp:2380-2510). One subscriber on the engine's
-		// event bus, fanning the typed events out to the publisher.
-		// The manifest cache feeds master_key resolution for
-		// pubValidation (NetworkOPs.cpp:2434-2438).
-		if consensusComponents.Engine != nil {
-			consensusComponents.Engine.Subscribe(&rpcEventBridge{
-				publisher: publisher,
-				manifests: consensusComponents.ValidatorManifests,
-				networkID: uint32(networkID),
-			})
-		}
-	}
-
-	ledgerService.SetSubmittedTxCallback(func(ev service.SubmittedTxEvent) {
-		publisher.PublishProposedTransaction(
-			buildProposedTxEvent(ev),
-			ev.AffectedAccounts,
-		)
-	})
-
-	serverStatus := newServerStatusPublisher(services, publisher)
-	ledgerService.SetServerStatusCallback(serverStatus.publish)
-	if feeTrack := ledgerService.FeeTrack(); feeTrack != nil {
-		feeTrack.SetOnChange(func() {
-			ledgerService.SignalServerStatusPublication(serverStatus.statusPublication(nil))
-		})
-	}
-	if consensusComponents != nil && consensusComponents.Adaptor != nil {
-		consensusComponents.Adaptor.SetOnOperatingModeChange(func(mode consensus.OperatingMode) {
-			ledgerService.SignalServerStatusPublication(serverStatus.modePublication(mode.String()))
-		})
-	}
-
-	// Wire up ledger service events to WebSocket broadcasts
-	ledgerService.SetEventSink(service.EventSinkFunc(func(event *service.LedgerAcceptedEvent) error {
-		if event == nil || event.LedgerInfo == nil {
-			return nil
-		}
-		transactions, err := buildValidatedTransactionPublications(
-			event.TransactionResults,
-			event,
-			uint32(networkID),
-		)
-		if err != nil {
-			return err
-		}
-
-		// Drive online-delete rotation off the validated-ledger advance. The
-		// callback fires from both the standalone accept path and the
-		// consensus SetValidatedLedger path, so the rotator sees every
-		// validated sequence. Notify never blocks.
-		if rotator != nil {
-			rotator.Notify(event.LedgerInfo.Sequence)
-		}
-
-		serverInfo := ledgerService.GetServerInfo()
-		ledgerCloseEvent := buildLedgerCloseEvent(event, serverInfo)
-		if ledgerCloseEvent == nil {
-			return fmt.Errorf("accepted ledger event %d has no source ledger", event.LedgerInfo.Sequence)
-		}
-		publisher.PublishLedgerClosed(ledgerCloseEvent)
-
-		publications, bookTransactions, failures := prepareAcceptedPublications(event, uint32(networkID))
-		for _, failure := range failures {
-			serverLog.Error("Skipping corrupt accepted transaction", "hash", upperHex(failure.hash[:]), "err", failure.err)
-		}
-		// pubBookChanges → book_changes aggregate stream
-		// (Subscribe.cpp:139-142 + NetworkOPs.cpp:3160-3174). Feed the
-		// already-closed ledger view directly from the event so a slow
-		// adapter store cannot drop the announce when the ledger isn't
-		// yet visible to GetLedgerBySequence.
-		bookView := newAcceptedLedgerView(event, nil)
-		payload := handlers.ComputeBookChangesFromTransactions(bookView, bookTransactions)
-		if data, err := json.Marshal(payload); err == nil {
-			wsServer.SubscriptionManager().BroadcastToStream(types.SubBookChanges, data, nil)
-		}
-
-		for _, publication := range publications {
-			publisher.PublishTransaction(publication.event, publication.projection.affectedAccounts)
-			pairs := acceptedOrderBookPairs(publication)
-			if len(pairs) != 0 {
-				publisher.PublishOrderBookChange(publication.event, pairs)
-			}
-		}
-
-		serverStatus.publish(nil)
-
-		// Update persistent path_find sessions on ledger close
-		wsServer.UpdatePathFindSessions(func() (types.LedgerStateView, error) {
-			return services.Ledger.GetClosedLedgerView()
-		})
-
-		serverLog.Debug("Broadcasted ledger",
-			"sequence", event.LedgerInfo.Sequence,
-			"txs", len(event.TransactionResults),
-		)
-		return nil
-	}))
-
-	if !standalone && appConfig.Watchdog.IsEnabled() {
-		wdCfg := watchdog.ConfigFromSeconds(
-			appConfig.Watchdog.WarnSecondsResolved(),
-			appConfig.Watchdog.FatalSecondsResolved(),
-			appConfig.Watchdog.AbortSecondsResolved(),
-		)
-		stallWatchdog = watchdog.New(wdCfg, nil)
-		if consensusComponents != nil {
-			if sp, ok := consensusComponents.Engine.(stallPinger); ok {
-				sp.SetStallPing(stallWatchdog.Register("consensus"))
-			}
-		}
-	}
-
-	shutdownCh := make(chan struct{}, 1)
-	services.SetShutdownFunc(func() {
-		serverLog.Info("Shutdown requested via RPC stop command")
-		select {
-		case shutdownCh <- struct{}{}:
-		default:
-		}
-	})
-
-	if err := context.Cause(ctx); err != nil {
-		return err
-	}
-	transports, err = bindRPCTransports(
+) error {
+	return newNodeRuntime(
 		ctx,
-		serverLog,
 		appConfig,
-		httpServer,
-		wsServer,
-		ledgerService,
-		systemListen,
-	)
-	if err != nil {
-		return err
-	}
-	if err := context.Cause(ctx); err != nil {
-		return err
-	}
-	if consensusComponents != nil {
-		if err := consensusComponents.Start(ctx); err != nil {
-			return fmt.Errorf("start consensus components: %w", err)
-		}
-	}
-	if err := context.Cause(ctx); err != nil {
-		return err
-	}
-	if err := transports.serve(serverLog); err != nil {
-		return err
-	}
-	listenerErrCh := transports.errors
-	if transports.grpc != nil {
-		serverLog.Info("gRPC server started", "name", transports.grpc.name, "addr", transports.grpc.address)
-	}
-
-	// Arm the out-of-band stall watchdog now that the server is up and
-	// servicing its event loops. Mirrors rippled arming activateStallDetector
-	// only at full start, and only outside standalone (ApplicationImp::run:
-	// guarded by !config_->standalone()). Standalone closes ledgers solely on
-	// the ledger_accept RPC, so an idle node produces no heartbeat and would
-	// otherwise self-abort; consensus mode drives a periodic heartbeat. The
-	// watchdog runs on its own goroutine and aborts the process if a monitored
-	// loop wedges, so a deadlocked node screams and can be restarted instead of
-	// going quiet.
-	if stallWatchdog != nil {
-		wdCtx, cancelWatchdog := context.WithCancel(ctx)
-		defer cancelWatchdog()
-		go stallWatchdog.Run(wdCtx)
-		serverLog.Info("Stall watchdog armed",
-			"warn_s", appConfig.Watchdog.WarnSecondsResolved(),
-			"fatal_s", appConfig.Watchdog.FatalSecondsResolved(),
-			"abort_s", appConfig.Watchdog.AbortSecondsResolved(),
-		)
-	}
-
-	var componentErrCh <-chan error
-	if consensusComponents != nil {
-		componentErrCh = consensusComponents.Errors()
-	}
-	if options.Ready != nil {
-		options.Ready()
-	}
-	return waitForShutdown(
-		ctx,
-		serverLog,
-		options.Reload,
-		shutdownCh,
-		listenerErrCh,
-		componentErrCh,
-		ledgerService.Errors(),
-		consensusComponents,
 		configPath,
-	)
+		standalone,
+		startup,
+		rootLogger,
+		serverLog,
+		options,
+	).run()
 }
 
 func configuredLedgerLoadFees(appConfig *config.Config) drops.Fees {
@@ -1408,28 +433,20 @@ func pebbleStoreOptions(n config.NodeDBConfig) (kvpebble.Options, error) {
 // (minutes) onto the node-object cache parameters, substituting the
 // built-in defaults for unset (zero) values.
 func nodeStoreCacheParams(n config.NodeDBConfig, nodeSize string) (int, time.Duration) {
-	profiles := map[string]struct {
-		size int
-		age  time.Duration
-	}{
-		"tiny":   {262_144, 30 * time.Minute},
-		"small":  {524_288, 60 * time.Minute},
-		"medium": {2_097_152, 90 * time.Minute},
-		"large":  {4_194_304, 120 * time.Minute},
-		"huge":   {8_388_608, 900 * time.Minute},
+	size, age := defaultNodeCacheSize, defaultNodeCacheAge
+	switch nodeSize {
+	case "tiny":
+		size, age = 262_144, 30*time.Minute
+	case "small":
+		size, age = 524_288, 60*time.Minute
+	case "large":
+		size, age = 4_194_304, 120*time.Minute
+	case "huge":
+		size, age = 8_388_608, 900*time.Minute
 	}
-	if nodeSize == "" {
-		nodeSize = "medium"
-	}
-	profile, ok := profiles[nodeSize]
-	if !ok {
-		profile = profiles["medium"]
-	}
-	size := profile.size
 	if n.CacheSize > 0 {
 		size = n.CacheSize
 	}
-	age := profile.age
 	if n.CacheAge > 0 {
 		age = time.Duration(n.CacheAge) * time.Minute
 	}
@@ -1614,326 +631,8 @@ func applyValidatorReloadContextWithGate(
 	)
 }
 
-const nodeShutdownTimeout = 30 * time.Second
-
-const (
-	transportShutdownGrace = 5 * time.Second
-	producerShutdownGrace  = 5 * time.Second
-	serviceShutdownGrace   = 10 * time.Second
-	storeShutdownGrace     = 5 * time.Second
-)
-
-func doShutdown(
-	transports *boundRPCTransports,
-	wsServer *rpc.WebSocketServer,
-	ledgerService *service.Service,
-	ledgerCleaner *cleaner.Cleaner,
-	consensusComponents *adaptor.Components,
-	rotator *shamapstore.Rotator,
-	kvDB nodestore.Database,
-	repoManager relationaldb.RepositoryManager,
-	logger xrpllog.Logger,
-) error {
-	return doShutdownWithin(
-		nodeShutdownTimeout,
-		transports,
-		wsServer,
-		ledgerService,
-		ledgerCleaner,
-		consensusComponents,
-		rotator,
-		kvDB,
-		repoManager,
-		logger,
-	)
-}
-
-func doShutdownWithin(
-	timeout time.Duration,
-	transports *boundRPCTransports,
-	wsServer *rpc.WebSocketServer,
-	ledgerService *service.Service,
-	ledgerCleaner *cleaner.Cleaner,
-	consensusComponents *adaptor.Components,
-	rotator *shamapstore.Rotator,
-	kvDB nodestore.Database,
-	repoManager relationaldb.RepositoryManager,
-	logger xrpllog.Logger,
-) error {
-	ctx, cancel := context.WithTimeout(context.Background(), timeout)
-	defer cancel()
-	var errs []error
-
-	transportsStopped, err := shutdownTransports(ctx, transports, wsServer, logger)
-	if err != nil {
-		errs = append(errs, err)
-	}
-	if !transportsStopped {
-		errs = append(errs, errors.New("shutdown incomplete: dependencies left running because transport handlers did not stop"))
-		return errors.Join(errs...)
-	}
-
-	producersStopped := true
-	if rotator != nil {
-		completed, err := runShutdownPhaseWithin(ctx, producerShutdownGrace, "stop online delete rotator", func() error {
-			rotator.Stop()
-			return nil
-		})
-		producersStopped = producersStopped && completed
-		if err != nil {
-			errs = append(errs, err)
-		} else {
-			logger.Info("Online delete rotator stopped")
-		}
-	}
-	if ledgerCleaner != nil {
-		completed, err := runShutdownPhaseWithin(ctx, producerShutdownGrace, "stop ledger cleaner", func() error {
-			ledgerCleaner.Stop()
-			return nil
-		})
-		producersStopped = producersStopped && completed
-		if err != nil {
-			errs = append(errs, err)
-		} else {
-			logger.Info("Ledger cleaner stopped")
-		}
-	}
-	if consensusComponents != nil {
-		completed, err := runShutdownPhaseWithin(ctx, producerShutdownGrace, "stop consensus components", consensusComponents.Stop)
-		producersStopped = producersStopped && completed
-		if err != nil {
-			errs = append(errs, err)
-			if completed && consensusComponents.Archive != nil {
-				archiveCtx, cancelArchive := context.WithTimeout(context.Background(), producerShutdownGrace)
-				terminalErr := consensusComponents.Archive.Close(archiveCtx)
-				cancelArchive()
-				if terminalErr != nil && !errors.Is(err, terminalErr) {
-					errs = append(errs, fmt.Errorf("close validation archive: %w", terminalErr))
-					logger.Warn("Validation archive terminal shutdown failed", "err", terminalErr)
-				}
-			}
-		}
-		if consensusComponents.Archive != nil {
-			health := consensusComponents.Archive.Health()
-			if !health.Healthy {
-				logger.Warn("Validation archive unhealthy at shutdown",
-					"enqueued", health.Enqueued,
-					"overload_dropped", health.OverloadDropped,
-					"closed_dropped", health.ClosedDropped,
-					"malformed_dropped", health.MalformedDropped,
-					"persistence_dropped", health.PersistenceDropped,
-					"write_failures", health.WriteFailures,
-					"retention_failures", health.RetentionFailures,
-					"last_error", health.LastError,
-				)
-			}
-		}
-		if completed {
-			logger.Info("Consensus components stopped")
-		}
-	}
-
-	if !producersStopped {
-		errs = append(errs, errors.New("shutdown incomplete: stores left open because producers did not stop"))
-		return errors.Join(errs...)
-	}
-
-	serviceStopped := true
-	if ledgerService != nil {
-		completed, err := runShutdownPhaseWithin(ctx, serviceShutdownGrace, "stop ledger service", func() error {
-			ledgerService.Stop()
-			return nil
-		})
-		serviceStopped = completed
-		if err != nil {
-			errs = append(errs, err)
-		} else {
-			logger.Info("Ledger service persistence drained")
-		}
-	}
-	if !serviceStopped {
-		errs = append(errs, errors.New("shutdown incomplete: stores left open because ledger service did not stop"))
-		return errors.Join(errs...)
-	}
-
-	if kvDB != nil {
-		_, err := runShutdownPhaseWithin(ctx, storeShutdownGrace, "close node store", kvDB.Close)
-		if err != nil {
-			errs = append(errs, err)
-		}
-	}
-	if repoManager != nil {
-		_, err := runShutdownPhaseWithin(ctx, storeShutdownGrace, "close relational database", repoManager.Close)
-		if err != nil {
-			errs = append(errs, err)
-		}
-	}
-
-	shutdownErr := errors.Join(errs...)
-	if shutdownErr != nil {
-		logger.Warn("Shutdown completed with errors", "err", shutdownErr)
-	} else {
-		logger.Info("Shutdown complete")
-	}
-	return shutdownErr
-}
-
-func runShutdownPhase(ctx context.Context, name string, stop func() error) (bool, error) {
-	if err := context.Cause(ctx); err != nil {
-		return false, fmt.Errorf("%s: %w", name, err)
-	}
-	result := make(chan error, 1)
-	go func() { result <- stop() }()
-	select {
-	case err := <-result:
-		if err != nil {
-			return true, fmt.Errorf("%s: %w", name, err)
-		}
-		return true, nil
-	case <-ctx.Done():
-		return false, fmt.Errorf("%s: %w", name, context.Cause(ctx))
-	}
-}
-
-func runShutdownPhaseWithin(
-	ctx context.Context,
-	timeout time.Duration,
-	name string,
-	stop func() error,
-) (bool, error) {
-	phaseCtx, cancel := context.WithTimeout(ctx, timeout)
-	defer cancel()
-	return runShutdownPhase(phaseCtx, name, stop)
-}
-
-func shutdownTransports(
-	ctx context.Context,
-	transports *boundRPCTransports,
-	wsServer *rpc.WebSocketServer,
-	logger xrpllog.Logger,
-) (bool, error) {
-	var errs []error
-	var servers []*http.Server
-	if transports != nil {
-		transports.stopRequests()
-		for _, bound := range transports.http {
-			servers = append(servers, bound.server)
-		}
-		for _, bound := range transports.ws {
-			servers = append(servers, bound.server)
-		}
-	}
-
-	type shutdownResult struct {
-		name string
-		err  error
-	}
-	graceCtx, cancelGrace := context.WithTimeout(ctx, transportShutdownGrace)
-	defer cancelGrace()
-
-	resultCount := len(servers)
-	if wsServer != nil {
-		resultCount++
-	}
-	if transports != nil && transports.grpc != nil {
-		resultCount++
-	}
-	results := make(chan shutdownResult, resultCount)
-	for _, server := range servers {
-		go func() {
-			results <- shutdownResult{
-				name: "drain HTTP server " + server.Addr,
-				err:  server.Shutdown(graceCtx),
-			}
-		}()
-	}
-	if wsServer != nil {
-		go func() {
-			results <- shutdownResult{name: "close WebSocket sessions", err: wsServer.Close(graceCtx)}
-		}()
-	}
-	if transports != nil && transports.grpc != nil {
-		logger.Info("Draining gRPC connections...")
-		go func() {
-			transports.grpc.server.GracefulStop()
-			results <- shutdownResult{name: "drain gRPC server"}
-		}()
-	}
-
-	graceful := true
-	for remaining := resultCount; remaining > 0; remaining-- {
-		select {
-		case result := <-results:
-			if result.err != nil {
-				graceful = false
-				errs = append(errs, fmt.Errorf("%s: %w", result.name, result.err))
-			}
-		case <-graceCtx.Done():
-			graceful = false
-			errs = append(errs, fmt.Errorf("drain transports: %w", context.Cause(graceCtx)))
-			remaining = 0
-		}
-	}
-
-	if !graceful {
-		for _, server := range servers {
-			if err := server.Close(); err != nil {
-				errs = append(errs, fmt.Errorf("force close HTTP server %s: %w", server.Addr, err))
-			}
-		}
-		if transports != nil && transports.grpc != nil {
-			transports.grpc.server.Stop()
-		}
-	}
-	if transports != nil {
-		if err := transports.closeListeners(); err != nil {
-			errs = append(errs, fmt.Errorf("close transport listeners: %w", err))
-		}
-	}
-
-	type joinResult struct {
-		name string
-		err  error
-	}
-	joinCount := 0
-	joined := make(chan joinResult, 3)
-	if transports != nil {
-		joinCount += 2
-		go func() {
-			transports.wait()
-			joined <- joinResult{name: "join transport servers"}
-		}()
-		go func() {
-			transports.waitRequests()
-			joined <- joinResult{name: "join transport handlers"}
-		}()
-	}
-	if wsServer != nil {
-		joinCount++
-		go func() {
-			joined <- joinResult{name: "join WebSocket sessions", err: wsServer.Close(ctx)}
-		}()
-	}
-	complete := true
-	for remaining := joinCount; remaining > 0; remaining-- {
-		select {
-		case result := <-joined:
-			if result.err != nil {
-				complete = false
-				errs = append(errs, fmt.Errorf("%s: %w", result.name, result.err))
-			}
-		case <-ctx.Done():
-			complete = false
-			errs = append(errs, fmt.Errorf("join transports: %w", context.Cause(ctx)))
-			remaining = 0
-		}
-	}
-	return complete, errors.Join(errs...)
-}
-
-// ledgerCleanerSource adapts the ledger service + node store to the
-// cleaner.LedgerSource interface the ledger-integrity verifier consumes.
-
+// buildTable builds the live amendment table from configured votes and overlays
+// any votes persisted by the relational repository.
 func buildTable(ctx context.Context, cfg config.AmendmentsConfig, repo relationaldb.RepositoryManager, log xrpllog.Logger) *amendment.Table {
 	t := amendment.NewTable()
 	for _, name := range cfg.Upvote {
